scripts/requirements.txt
yfinance>=0.2.0
supabase>=2.0.0
yahoo_fin>=0.8.9
pandas>=1.5.0
requests>=2.28.0
schedule>=1.2.0
python-dotenv
lxml
pytz
requests-html
